requirements.txt
alembic==1.4.3
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
